modify login page

+++ OneblogiOS/EmojiPanelVC.m
@@ -111,7 +111,7 @@
             emojiStr = Utils.emojiDict[@(emojiNum).stringValue];
             emojiImageName = [emojiStr stringByTrimmingCharactersInSet:[NSCharacterSet characterSetWithCharactersInString:@":"]];
         } else {
-            emojiStr = [NSString stringWithFormat:@"[%ld]", emojiNum - 1];
+            emojiStr = [NSString stringWithFormat:@"[%d]", (int)emojiNum - 1];
             emojiImageName = [NSString stringWithFormat:@"%03ld", (long)emojiNum];
         }
